Moves GetTestStateScheme to util/testhelpers/env
